we can express the generic type in a single case of enum by using an alias
Uff! Without this we'd be in trouble!
This is super important because messages can be numerous,
and many of them have completely unrelated generic things (ActorRef<A>, ActorRef<B>, or Behavior<C>),
so if we had to put all those A B C into the enum up there passing the enum around in type signature would
quickly become insane, as it would become: `ActorRef<Message<A, B, C>>` and those normally are not letters,
but very long names of messages; so you can see we'd be in trouble there.

Thank goodness for bouncing ideas a bit with Johannes and I remembered that we can use aliases, which solevs it