Working mailbox scheduling with tests; system/user, sadly with Lock
Implementation generally is prepared to perform all mailbox operations
in a wait-free way. It is optimized for "common thing" which is an
actor, which is scheduled to receive more messages.

I believe we need a guarantee that the write to the mailbox queues
themselfes is never visible to another thread until the write to the
mailbox status is made.