Fix race condition in Mailbox when enqueueing system messages #138
We now use 2 bits for keeping track of system messages. LSF is for signaling that messages are in the queue, the next highest bit is to signal that we are currently processing them. Also reversed the order in which we enqueue system messages and set the bit. In this case instead of not processing the messages, worst case we schedule without having any messages.