=serialize #418 Codable serialization infra for ReceivesSystemMessages
While this is not the most popular of use cases I think, as it is
confined to our library itself sending those messages, it is good
to make sure we would be able to handle such message if we needed to.

Automatic Codable derivation does not entirely work, as the types
involved are internal, and shall remain so -- as users should not touch
the internal `ReceivesSystemMessages` types, after all.