Fix segfault in error handling #559
This was caused by a use after free of the message pointer. The way we stored the message led to it being freed when longjmp'ing back from the signal handler. This patch uses the out pointer that gets passed in to capture the message that caused a failure, which is heap allocated