+act Basic WorkerPool, good enough for Simple/someone cases
* +act #642 Basic WorkerPool implementation, receptionist powered

Very basic implementation which does not do fancy nor configurable
balancing (not even stable in face of member changes). But for pure
stateless workers i.e. handling http requests as-if a thread pool this
is already somewhat useful.

The general idea is that a pool is always powered by a receptionist key.
Implementations using this can ask for a key or create the pool
themselves.

This currently takes the hit of being an actor, but could be made
not-an-actor so the forwarding would only take one hit I believe.
Currently this already is quite good enough so not spending time on the
non-actor impl.

We may revisit naming at some point... currently all uses are "workers"
but technically this is just "actor pool". I'm unsure about adopting
Akka's naming here, to be honest it always was a bit confusing, so
perhaps we can ride on the receptionist _always_ and only add abilities
to "lock in members" or similar :thinking: