Notify ActorOwned CRDT instance after successful registration
Motivation:
An `ActorOwned` CRDT instance is not ready for use until registration with the local replicator is complete. Ideally an actor would remain in a non-ready state until all of its `ActorOwned`s are ready.

Modifications:
- Add `ready` to `DataOwnerMessage`, which gets sent by the local replicator upon successful `ActorOwned` registration.
- Allow owner to define custom `onReady` callback.

Result:
An actor can wait for `ready` signal before mutating a CRDT instance.

This is related to https://github.com/apple/swift-distributed-actors/issues/14.