=sup #31 Implements .escalate supervision scheme
== Motivation

.escalate makes failures be escalated to parent upon failure,
these are escalated more if the parent also does .escalate
automatically.

If one wants to isolate the failure, i.e. "stop it from bubbling up",
keeping the supervision as .stop does the trick -- the actor which is
.stop supervising stops, rather than escalates the issue.

Watching is inherently "same-ish" to this -- watching is a superset of
supervision escalate. A parent always is notified about child
termination. Only watching a child though makes the death pact active.

Alternatively, one may want to supervise the spawned child with
.escalate, meaning that any failures the child does we shoudl get...
but we will NOT get its .stop signals (!) which watching WOULD give us.