The cluster events are dead letters when cluster is not started, thus dont subscribe there to avoid warning messages
**Motivation:**

Less noise when running the system in non clustered mode and the CRDT
replicator boots up

**Modifications:**

Only subscribe when the cluster is enabled.
Could be worked around more nicely to only start the replicator when we
have clustering enabled?

**Result:**

Avoids: `2019-10-31T11:01:54+0900 info: deadLetter=1
[ActorMessageFloodingBenchmarks][EventStream.swift:41][/dead/letters]
Dead letter:
[subscribe(ActorRef<ClusterEvent>(/system/replicator/$sub-DistributedActors.ClusterEvent-y))]:DistributedActors.EventStreamShell.Message<DistributedActors.ClusterEvent>
was not delivered to ["/dead/letters"].`