Allow for inspecting of returned types in Actorable, act on behaviors
**Motivation:**

Two reasons:

- firstly, we currently have no other way of stopping an actor; so we'd
need to return .stop; Though this will change as we'll do context.stop()
I think as we change "styles"; note at the same time
- for other reasons we may want to allow returning behaviors, as we have
the suspending mechanisms which we cannot model otherwise at least not
yet.

And secondly:

- We'll want to allow by convention, replacing a replyTo: ActorRef with
a method that returns a value. The worrying bit there is the blocking
story, but once async/await lands that will be addressed.

**Modifications:**

Parse and keep the return type of actorable functions;
Emit a `return instance.invoke()` if a behavior is being returned.

**Result:**

- Ability to use behaviors from an actorable

Exclude generated sources from formatting checker for the time being