Must not touch EL based APIs after shutting down the ELG
The newly added repeated task that does tombstone flushing was cancelled AFTER the ELG shutdown, leading to the  `ERROR: Cannot schedule tasks on an EventLoop that has already shut down. This will be upgraded to a forced crash in future SwiftNIO versions.` error explained in Don't schedule futures on terminated EventLoop #922.

We must not touch ELG related APIs after shutting down; A cancel seems to create a future as well, so it was causing issues.