Avoid errors in old toolchains, add to-be-removed convenience keyword
It was reported that beta-2 toolchain errors out on the missing `convenience` in this init. 

This is actually fixed by https://github.com/apple/swift-evolution/blob/main/proposals/0327-actor-initializers.md which @kavon implemented, so as soon as we get a stable Swift which does not need the convenience keyword we can drop it.

Adding for now to make life of early adopters easier.