Fixed a crash on Swift 6.0+ where Integer deserialization was assuming aligned memory (#1190)
### Motivation:

The integer deserializer was assuming that the input raw memory was
correctly aligned for the integer type to be deserialized. In the test
that was true on some platforms, but not others.

This caused a crash in the unit test.

### Modifications:

Use `loadUnaligned` instead of `load` when loading integers.

### Result:

Fixes https://github.com/apple/swift-distributed-actors/issues/1179.