Fix the Samples project so that it compiles and runs. (#1205)
### Motivation:

Currently, the sample project doesn't compile because
`swift-cluster-membership` requires macos 14.0.

```console
error: the executable 'SampleDiningPhilosophers' requires macos 13.0, but depends on the product 'DistributedCluster' which requires macos 14.0; consider changing the executable 'SampleDiningPhilosophers' to require macos 14.0 or later, or the product 'DistributedCluster' to require macos 13.0 or earlier.
```

### Modifications:

Update the `Package.swift` file in `/Samples` to require macos 14 and
bump to Swift 5.9 to support running in distributed mode.

Updated README to include `swift` in the run commands.

### Result:

`Samples` project now compiles and runs.