Workaround SwiftPM API breakage checker limitation (#1211)
### Motivation

In order to address a hanging integration test on newer Swift versions,
#1208 added a target dependency on the executable used in the multi-node
test plugin and stopped shelling out to `swift run`. This addressed the
hang, but it has surfaced another issue: that SwiftPM's API breakage
checker has issues with some packages that contain plugins, and now that
pipeline is reproducibly resulting in false positives.

### Modifications

This patch removes that dependency in the package manifest and reverts
to running it from within the plugin. There was already code in the
plugin to build the package and locate the required binary in the built
artifacts, which I hadn't noticed in the previous patch, so the use of
`swift run` seemed unnecessary. This patch just executes that found
executable without `swift run`, which (hopefully) resolves both issues.

### Result:

- Integration test should run to completion on all Swift versions (not
hang on 6+)
- API breakage pipeline should pass.