Fix compilation on platforms without Foundation.Process (#1209)
### Motivation:

There's an open PR to add CI for Apple platforms. This is currently
failing since it tries to build for all Apple platforms, not just macOS.
This is because there are targets that provide support for testing that
rely on `Foundation.process`, which is unavailable on non-macOS Apple
platforms.

### Modifications:

Guard the use of `Foundation.Process` with an `#if canImport`. 

### Result:

Package builds for non-macOS Apple platforms: iOS, tvOS, watchOS, etc.